At installation, unregister example suites for the admin user.
The 'create-cdb' admin command registers the example suites in the admin
user's private suite db then exports them to the central db. It now
unregisters them in the private db after the export so that even the
admin user has to import *copies* of the example suites. This should
prevent the admin user from making accidental changes to the originals.