Pseudo version number for running cylc within the repository.
Uses 'git describe', which describes the current repository state in
terms of the most recent version tag, the number of commits beyond it,
and the ID of the latest commit. This means User Guides generated from
the repository, as opposed to when generating a new release, will have
a sensible pseudo version tag.