Converted to new task retry method, delays not implemented yet.
Command scripting was a list of (potentially) multi line strings; now
it is a single multi line string that is repeatedly invoked, but with an
environmental variable CYLC_TASK_TRY_NUMBER to allow different behaviour.
The number of retries is determined by a new list item, "retry delays,
"retry delays".