If a remote task or owner is defined use ssh to submit.
Previously we checked that remote host != localhost and owner != suite
host before using ssh job submission. But it is sometimes useful to
run local jobs via ssh, at least for test purposes, as it disconnects
the task from the suite environment which is otherwise difficult to do
for local tasks (even the 'at' scheduler copies the job submission
environment for the submitted job).