Use a bash login shell for the remote command
Adding the --login flag to bash automatically sources the profile files
for the remote machine, avoiding the need for checking for them in the
ssh command.

Whether or not to use a login shell is now enabled by a boolean within
execute(), which will be changed to a config file setting.