Add login shell option for remote tasks
Add a new suite.rc option:

    [runtime]->[[_NAME_]]->[[[remote]]]->use login shell

Enabling this option results in remote commands being run in a login
shell, so that a user's ~/.profile settings are sourced. It defaults to
True.

Disabling this setting may be useful for facilities that wish to limit
the creation of unattended login shells for security reasons.

With the option enabled remote commands look like

    ssh user@host "bash --login cylc function"

with it disabled they look like

    ssh user@host "cylc function"

Also provided is the command line options --login and --no-login, which
enable & disable this functionality respectively from the command line.