Don't recreate and reconnect the suite identifier on reload.
To reconnect to Pyro, first we'd have to disconnect the original; but
that's unnecessary because the suite ID does not change on reloading
the suite definition. This bug was revealed when attempted disconnection
failed in the suite shutdown code, because Pyro no longer referenced the
right identifier object. This is also a bug in the current master branch
but it has no effect there as we don't explicitly disconnect from Pyro
on shutdown.