htlatex command line to fix navigation links.
Most of the page header and footer navigation links (next, prev, tail,
etc.) inserted by htlatex in the multi-page version of the User Guide
were wrong. I found a comment on the web suggesting that this happens
to documents with footnotes unless the 'fn-in' option is used (which
puts foonotes on their parent page, I think). This did fix the problem.