Run the cylc server in daemon mode; closes #84.
Daemonizes the cylc server process just after connecting to the suite
Pyro port, so that we can check the suite is not already running before
daemonizing and redirecting output to log files.  Some useful info is
printed to stdout before daemonizing. The --debug option causes cylc to
run in non-daemon mode.