Better use of flag to stimulate dependency negotiation.
Incoming messages that change task outputs set a flag to signal that
dependency negotation is required. The flag should only be turned
off again after it has been used, else dependency negotation could be
missed if the flag gets turned on then off in quick succession. This
can result in a stalled suite that requires a "nudge" to get it going.