Fixed T-N cycle offset bug for n > 24 (HoursOfTheDay cycling)
Example:
[[[0]]]
    graph = "foo[T-36] = bar"
This graph results in bar triggering off foo at T-36; but it is also
used to infer that one of the valid cycle hours of foo is 0z minus 36,
or 18z. Conversion of T-N back to the 0-23 range was broken for N > 24
with the result that foo could sometimes spawn into the wrong cycle.