cylc/cylc#861: fix cylc validate hyphenated family trigger.
This fixes the problem of checking the family 'FOO'
against the line 'FOO-BAR:succeed-all => baz'. This
fails validation because the '-' looks like a line
break and therefore looks like a bare 'FOO' family
trigger. This change fixes the problem by reverse
sorting the family names before checking each one.
As a shorter subset of a name sorts before a name,
and we're reverse-sorting, this means that
'FOO-BAR' is expanded before 'FOO' gets to it.