Respond to lost connection better in gcylc suite views.
View-clearing code was called repeatedly instead of just once when not
connected to the suite. In the graph view this involves a graphviz
subprocess, repeated calls to which could result in hundreds of error
dialogs if the subprocess was failing for some reason.

Also tidies a bit of old code, splits the dot- and tree-view updater
classes into separate files like the graph-view updater, and renames a
couple of files more sensibly.