Support loadleveler job pre-empt/vacation
Handle `SIGUSR1` for loadleveler jobs with `restart=yes` directive.
On receiving a `SIGUSR1` signal, a job script will remove its status
file, and send a message to the suite to indicate that it has been
vacated. The server will reset the status of the task (job) to
`submitted`. This change resolve cylc/cylc#867.