Pass $CYLC_SUITE_RUN_DIR to ssh-invoked message commands.
With "task communication method = ssh messaging" this bug caused the
suite env file to be ignored by messaging commands re-invoked by ssh on
the suite host, because the file path defaults wrongly to
"./cylc-suite-env" if the suite run directory is not known.  Although
this is a bug, it would have no effect under normal circumstances
because the suite env file on the task host will still determine the
suite variables passed to the environment of the re-invoked message
command on the suite host.