cylc gui: fix gtk warning for toolbar.
GTK versions 2.20+ produce a warning when cylc gui
is loaded, due to the use of a gtk.HBox in the
toolbar (in a gtk.ToolButton). This seems OK from
the documentation (and works fine) but it seems that
the GTK code is set up to only be happy with a
gtk.Misc object - e.g. a gtk.Image.

We used a gtk.HBox for the iconwidget in the
layout toolbutton (toggle side-by-side) so that we
could have a left and a right arrow - this now
has to be changed. Given that we can only use one
of these (only a single gtk.Image), I've moved this
to the View 2 toolbar to make it clearer.