Restructure suite run time log/job/ and work/
Directory structure now looks like:
* `log/job/CYCLE/TASK/{01,02,...,NN}/{job,job.out,job.err,job.status}`
* `work/CYCLE/TASK/`

which should make it easier to archive the log for the whole cycle or to
housekeep all the work files of a given cycle.

`log/job/CYCLE/TASK/NN` is a symbolic link that points to the job log
directory of the latest submit for `TASK.CYCLE`.