#119: fix initial cycle point handling in restart.
This fixes picking up the correct initial cycle point from a state file
and also clarifies some of the naming - it gets confusing when we are
either storing cycle points as strings or as objects, or moving between
the two.

This fixes a problem with the runahead pool change where tasks
beyond the stop point are no longer present in the state file on
stop-point-based suite shutdown. The restart and state-dump tests
need to be changed to reflect the state file alteration.

This also adds an initial cycle point option for "cylc validate", for
use when an initial cycle point has not been specified in the suite.rc
file.