Fix unbounded final-cycle-point sequences.
Sequences based on the final cycle point are created by
counting backwards from it. Sequences that also have a
fixed number of repetitions are bounded and also have
well defined start points. This means that they can be
reversed in the isodatetime library, so that they start
counting from their own start point.

However, sequences based on the final cycle point with
no repetitions specified are unbounded, and don't have
a start point in the isodatetime library. This means that
they generate points counting backwards from their end
point, which is in the wrong order for cylc.

This change generates a start point for these unbounded
final cycle point sequences so that the result is points
in the right order. We can just guess the start point
by iterating back from the end point until we reach
the initial cycle point area.