#940: cylc gui: optimise tree view updating.
The previous behaviour of the tree view on update was to clear the tree
model then rebuild it. This caused a noticeable blip for largish suites
and made it difficult to focus on a particular row.

The new behaviour is to update the tree view in place if possible (if
there has been no structural changes), or to carefully reconstruct it
when there are structural changes. This is done piece by piece rather
than clearing and rebuilding from scratch.

The time taken to update the tree is much reduced in the no-structural
case, but slightly more when a reconstruction is necessary (only when
updates introduce new tasks, remove them, change family structuring or
change the grouping flag). In all cases it is no longer as jumpy.