Handle spaces in CYLC_DIR when checking the version number in a git repo.
The current method uses subprocess, which passes the strings unescaped into
a shell - when CYLC_DIR contains spaces, they are not properly escaped which
causes a "Failed to get version number" error because attempt to run the
get-repo-version script stops at the first space in the path.

Adding quotes around the path when passing the command to the shell solves
the issue.