Fix pre-existing cylc gui graph bug for new runs.
This fixes a blank cylc gui Graph View when a suite is re-run
while the cylc gui is still around from the previous run. The
root cause is the initial request for the last run's oldest and
newest cycle times, which is stored in the 'actual_first_point'
variable (indirectly). This means that the subsequent requests
for the correct cycle point range do not give valid graphs,
because they appear to be well before the actual first point.

I've also fixed a bug in the purely-base-node-cycle-point cropping,
which would not have redrawn and uncropped the graph if e.g. a node
was re-inserted.