Fixed some tests.
These changes are mostly to do with succeeded tasks being eliminated quickly
now.  In the case of restart/01-broadcast.t this results in a broadcast setting
expiring earlier because the min cycle point in the suite changes. For
restart/14-back-compat-restart.t, after the restart both upstream tasks in an
OR trigger now get reference-logged as triggering the downstream task because
they both succeeded before the shutdown.