cylc gui: preserve edges on node filter out.
Currently, when a node is filtered out, the incoming
and outgoing edges from it are discarded. This can split the
graph and can mislead the user.

This change replaces all connected groups of filtered-out nodes
with small placeholder special nodes.

It adds an extra option to cylc graph to turn on filtering for
test purposes.