cylc.task_proxy: more info in summary messages
Task summary now includes the submit number. This gets displayed with the
latest message in the `cylc gui` tree view, and in tool tips in other views.

On retry, the latest message includes the delay interval and the time after
which it will retry.

On initiation of job submit, poll and kill, log the submit number.