#294: remove get_suite_host sys.exit just-in-case
This removes the sys.exit from the get_suite_host function. This is only
called at the start of scheduler.py (which would be fine) and during task proxy
email event hooks (at which point the error could not occur, as the global
config is already parsed). This is really just a just-in-case change.