cylc trigger --edit: fix timestamp cmp
The old logic compares the localhost time with the timestamp of the job
file. A problem occurs when our desktops and our file system servers
have gone out of sync with time, with the file system servers being a
bit behind, so the old logic thinks that the job file is never updated.

The safer way to detect change is to look at the time stamp of the job
file before and after.