cylc auth: allow hash choice, SHA256 by default
This change allows a site or user configuration of the hashing
algorithm used in Pyro communication. MD5 is not really recommended for
HMAC use, and is actually disabled on some POSIX systems.

Multiple hash algorithms are now allowed for both client and daemon.
The daemon can accept authentication in any of the configured hash
algorithms, and the client will work through any of the configured hash
algorithms until one works.

The default is SHA256 followed by MD5 as a fallback and for scanning.
This allows MD5 (old) suites to work - it preserves
old-client/new-daemon and new-scan-client/old-daemons.

SHA1 can be used for older Python version sites that don't want MD5.