cylc.gui.tailer: fix SSH zombie tail
On signal to "ssh" client, a signal is sent to "sshd" on server.
However, "sshd" cannot send a signal to the "tail" command, because it
is not a terminal. This leaves the "tail" command in a zombie state.
Apparently, we can use "ssh -t" or "ssh -tt", but that just causes the
command to hang for some reason, (which hangs the GUI as well). A quick
hack here is to do something like "tail --pid=$PPID", so "tail" dies as
soon as PPID dies.