Fix validation error when graphviz is not available
Issue #1463 added the capability to fail cyclic dependencies when
validating a suite.  However, if the graphviz library is not available
the graphing_disabled flag is set, but not checked for the cyclic
dependency generation.  This results either an error with traceback
(for cylc 6.6.1) or a cryptic "global name 'graphing' is not defined'
error in 6.7.1 but the validation always fails.

This updates the cyclic dependence check to run only if the graphing
module was imported successfully.  If the graphing module is not
available, a warning message is printed out (even if the suite is
otherwise valid).