Add option for sorted task names in the text-based monitor.
The current align-column option is only useful for small suites, but for
larger suites using the monitor the current list of tasks is presented in
a non-intuitive order.  This changeset allows a command-line option to
sort the names of the task proxies (using standard Python string sort order)
before displaying in the text-based `cylc monitor`.