#1744: fix constant state dumps for queued tasks.
This fixes a problem where the state file is continually dumped
if the task pool contains a queued task.

This is due to the ready_to_run check in process_tasks in scheduler.py
returning True if any task is queued and has reached its start time.

The fix is to only update the state summary if we really think
something has changed. This is a good thing to be doing anyway.