cylc refresh: no nuke passphrase on title change
It was nuking the passphrase by doing an unregister and a new register
on suite title change. This is not necessary, and on running suites,
will cause subsequent client commands to fail.

Also modified cylc.registration to improve its pylint score.

The new test would time out before this change, because the refresh
command would reset the passphrase, which would make it impossible for
subsequent task jobs to report progress back to the suite.