Nudge insert
By moving the logic to process the command queue to an earlier position
in the main loop.

Allow an inserted task to be released from the runahead pool so that it
is listed as a waiting task by the GUI. Previously, it would be listed
as a runahead task, and a nudge would be needed to release it to the
pool.