Allow external triggers to use the cycle point.
While external triggers are handy to avoid polling, the current
specification does not allow any cycle point information to be used
in the external trigger string within a task definition that will be
matched against incoming trigger messages.

This change permits an external trigger to include the current cycle point
(specified as "$CYLC_TASK_CYCLE_POINT") as part of the special tasks
in the suite scheduling definition.  At runtime, the cycle point is used
to perform the match, which means that only a message containing the
correct cycle point will result in the external-trigger dependency being
fulfilled.