Fix suite info on daemonize
cylc run/restart command will now poll for suite log to be populated
with the start up line so it can get the information on PID and port number and print them out correctly to the user.

Incidentally, this also allows the suite to exit 1 if the suite does not
start up correctly due to suite configuration errors.