cylc run|restart: remove old contact file
On suite start up, if an old contact file is found, read host and
process string from the contact file and run a (ssh+)ps command to
determine if an old suite is still running or not. If it can confirm
that the old suite process does not exist, it will remove the old
contact file and continue with starting up the suite.  Otherwise, it
will print the old error message.