Improve stall logic in presence of clock trigger
Before this change, if the suite has any clock triggered tasks, the
suite will wait for all clock triggered times to reach before considering
the suite as stalled, even if the clock triggered tasks can never be
triggered due to dependencies issues.

With this change, the suite will treat waiting clock triggered tasks
with unsatisfied dependencies the same as other waiting tasks with
unsatisfied dependencies. However, it will consider a clock triggered
task as an *active* task if it has its dependencies satisfied but is
only waiting for the clock triggered time.

The change allows the stalled event to be raised as soon as the
following conditions are satisfied:
* There is no activity after one main loop.
* The task pool is left with only inactive tasks.
  * Active tasks are those that are not pending outputs of other tasks.
* The task pool has inactive tasks with unsatisfied dependencies.

Fix test UTC