Update timing query to use task states table.
Timing information requires obtaining the most recent submission
number of a succeeded task.  When this is done with the task
events table, it requires a subquery but using the task states
table (where only the most recent submit number is stored) allows
a much simpler join operation to be used instead.