Optimise time taken to get raw graph
The main change is the restructure of the variable
`SuiteConfig().edges`. It is now a dict. Each key is a unique sequence,
and each value is a set of `(left, right, suicide, conditional)` tuples.
This change allows `get_graph_raw` to look at each unique sequence once
instead of multiple times. There is also caching at various points to
reduce the number of expensive calls to regular expression matching and
date-time functionality.

Other minor changes include:
* Remove unused variables.
* Use compiled regular expressions where possible.
* Reduce usage of complex regular expressions.
* Remove edge and graphnode objects, which were not cheap to initialise.
* New GraphNodeParser object to manage and cache graph node parsing.
* The `recursive_replace` internal function is replaced with a loop at a
  level up.