Update signal for gscan menu dialog.
With the configure hosts dialog moved to a proper menu in Cylc 7,
the dialog would not display for some users.  Investigation showed
that the function to create and display the dialog was not called,
because the button press signal was not being registered.  Replacing
button press with "activate" signal resulted in the dialog displaying
properly.