Allow debug output to include shell sources
The current shell script will source files and redirect all output
to /dev/null - this works very nicely unless there is an error hiding
in one of those files (or, more specifically, in a users ~/.profile or
~/.bashrc or equivalent).

Create verbose output when sourcing other files if the debug flag is
enabled to give at least some information when trying to trace errors.
First choice was to check error codes, but I can't make that work with
sourcing files (rather than calling them as scripts).