gui: further connect schedule changes
Unify connection schedule for running and stopped suites, with maximum
update duration being PT15S. This is now OK for stopped suites because
the client will look for suite's `host:port` via contact file first, so
it should only hit the file system every 15S instead of the network.

Only display connect count down if the update duration is *at* the
maximum PT15S level. This is when a running suite that does not see much
update and when the suite has stopped.

Increase in update duration will now be a bit more gradual.