client: smarter logic with failed connection
If a stopped suite has left behind a contact file, the client will now
remove it on failed connection, and report that the suite has stopped.

Adjust general error message on detection of a contact file of an active
suite - more information on the old suite is and how to shut it down.

Run `fsync` on writing a contact file.