Fix duplicate index bug in timings report.
If a task succeeded, and was later re-added to the task pool and ran,
there would be multiple entries in the suite run database for the
succeeded task at the same cycle point.  This caused problems with
creating the Pandas pivot table to re-organize the data by task name
for the diagnostic plots as part of the report-timings command.
This fix adds a check for when this error will occur and uses a
different reshaping method that avoids the original error.

Also, per the original suggested change, this bumps the maximum
column width to avoid truncation of printed data frames (especially
important for long task names).

See #2509.