Updated graph_parser.py to fix validation errors
Currently, the graph parsing will not raise an error when there are
spaces between task names, as spaces get stripped out
`foo => bar baz` becomes `foo => barbaz`.
It can be captured if that task does not exist, but if there was a
task with that name, it could lead to some odd results.

There are also some other oddities in the validation which should not
be allowed to happen. For example, these validate although I don't
think that they should:
```
foo[-PT6H]<a> => bar
baz<remote::task>[-PT6H] => bar
```

The approach taken with this change is to make a master regex which
only accepts appropriate orders of patterns. This covers the space
bug, but also prevents the incorrect order or combinations being
specified. All problem lines are reported as per the current approach
(but moved into a method).

Not covered in the regex is checking whether the remote trigger is
only using task state triggers, and not using family triggers.