Updated graph_parser.py to fix validation errors
Currently, the graph parsing will not raise an error when there are
spaces between task names, as spaces get stripped out
`foo => bar baz` becomes `foo => barbaz`.
It can be captured if that task does not exist, but if there was a
task with that name, it could lead to some odd results.

This change takes a simple approach to stop the space separated task
validation problems, and also to prevent bad spaces in parameters and
around triggers. It will not capture all potential validation problems,
some of which are covered by later validation checks.

It was not possible to just do task names and to ignore paramters and
triggers, as they generally follow the same format.