Fix some tests in graph_parser which were giving false passes
The tests for order of trigger and parameter were correctly raising a
GraphParseError, but, not for the reason thought. They were raising
the error because parameters were not defined, not because the order
was being correctly identified. I have now added in the parameters,
and they do not pass anymore.