localhost background/at jobs: record host name
Ensure that localhost background/at jobs are recorded as running on the
host name of the current suite host, rather than just `localhost`. On
suite restart on a different suite host, this allows the restart logic
to correctly poll the status of the background/at jobs that may still be
running on the previous suite host.