Replace flaky tests/events/02-multi.t
According to the summary line, the test exists to prove that multiple
event handlers can be specified for a task event. A simple test with 2
echo commands + an inspection of the suite log should be sufficient for
testing this behaviour - but should be much more reliable.

The problem was that task event handlers are not guaranteed to invoke in
order. In a busy system, the old tests based on state manipulation
commands could be invoked or processed in the wrong order causing the
test to misbehave.