host appointer: improve handling of unexpected out
On SSH call with login shell, sites/users may have logic in profile
scripts writing to STDOUT. This causes the logic that expects JSON to
fail to load the output. This change attempts to drop all lines before
it finds the first open curly bracket `{`, which should be good enough
to denote the start of JSON content in most cases.