Fix tests/jobscript/19 and others
tests/jobscript/19: This test was very time dependent previously -
especially on the third run of the suite. This fix should make it much
more event dependent, and also ensure that result in the second run of
the suite can no longer interfere with the logic in the third run.

tests/shutdown/10: remove debug cat

tests/shutdown/18: fix background + wait logic