task remote init: fix strange tempfile issue
Users are experiencing issues on one of the platforms we deloy Cylc
where remote submission fails intermittently with a file not found error
complaining that a temporary file cannot be found. (The file is used for
piping service files via STDIN (to the SSH) command.)

In this change, we pass the temporary file handle directly instead of
passing the name of the temporary file to the process pool. This appears
to fix the issue. (Note: Before #2590, we were unable to pass the file
handle because the context needs to be serialised for the
multiprocessing.Pool. This is no longer a requirement in our own
subprocess pool, so it is now safer to pass the handle.)

Add basic unit tests for running command with STDIN in subprocess pool.
Rename SuiteProcPool to SubProcPool.