Targeted clean: Overhaul processing of glob results
This filters out all redundant matches, i.e. matches that are subpaths
of other matches (except for the std symlink dirs - we need to return
these as we need to follow their targets).

As was implemented in a previous commit, it still also filters out
subpaths of symlinks as we don't want to follow symlinks (again except
for the std symlink dirs).