job: don't use $HOME to source the job script
The archer2 machine does not have access to `$HOME` on the compute
nodes.

On the whole we have handled this use case ok, however, one sticking
point remains.

Because Cylc 8 sources the job script from the `cylc-run` directory this
is a problem for them:

```bash
. "$HOME/cylc-run/two/.service/etc/job.sh"
```

This PR uses `$0` to locate the job script rather than `$HOME` to avoid
this issue.