Forward arbitrary environment variables over SSH
Allow users to specify environment variables that get forwarded from the
submit host to the scheduler and task submission.

Define variables to forward in `global.cylc` like:
```
[platforms]
    [[localhost]]
        ssh forward environment variables = PROJECT, LUSTRE_DISK
```

This will add `PROJECT` and `LUSTRE_DISK` to the list of variables
exported in SSH commands to launch the scheduler on remote hosts (if
they have been set in the current environment).

Once they are available on the scheduler they can further be forwarded
to other platforms, where they may interact with the scheduler to
set a default project code or be used to set storage paths:
```
[platforms]
    [[mtu]]
        ssh forward environment variables = PROJECT, LUSTRE_DISK
        install target = mtu

[install]
    [[symlink dirs]]
        [[[mtu]]]
            share = $LUSTRE_DISK/$PROJECT/$USER
```