Add new 'RAW' variables for SHARE and WORK variables
CYLC_WORKFLOW_SHARE_DIR_RAW and CYLC_WORKFLOW_WORK_DIR_RAW added.
These are then used for 'mkdir' to ensure directories are created.
Currently, if SHARE_DIR and WORK_DIR are symlinks, and the dir
they point to does not exist, the dirs will not be created, and
tasks may fail. The new variables ensure the correct directories
are always created.