Use here-document for job script environment variables
This prevents unwanted quote stripping, etc. As of this commit it does
not yet handle expanding ~, which the original code does, and has a
comment explaining why.

The here-document is unquoted, so shell parameters are expanded within.
This matches the existing behaviour, but if it is desired to make them
more robust still it can be easily prevented by quoting the first
opaque_terminator with double quotes.