Log the username of CLI users
This passes the discovered CLI user, if available, and adds them with a
reference to who it might be. Uses getlogin by default, which grabs the
user based on the controlling terminal and if that fails uses getuser
which pulls from environment variables.

This passed couched in the terms of "possibly on behalf of" as this is
data from the client and so is not reliable.